requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.4.1
    # via streamlit
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.5.1
    # via streamlit
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
idna==3.10
    # via requests
importlib-resources==6.4.5
    # via
    #   jsonschema
    #   jsonschema-specifications
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.25.2
    # via altair
numpy==1.24.4
    # via
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.0.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==10.4.0
    # via streamlit
pkgutil-resolve-name==1.3.10
    # via jsonschema
protobuf==5.29.3
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.20.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.40.1
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   rich
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
watchdog==4.0.2
    # via streamlit
zipp==3.20.2
    # via importlib-resources
